LayoutTests/http/tests/inspector/search/search-in-resource.html

<html>
<head>
<script src="../inspector-test.js"></script>
<script src="../resource-tree/resource-tree-test.js"></script>
<script src="search-test.js"></script>
<script>
function test()
{
    // This file should not match search query.
    var text = "searchTest" + "UniqueString";
    InspectorTest.runAfterResourcesAreFinished(["search.js", "search.css", "search.html"], step2);

    function step2()
    {
        var resource = WebInspector.resourceForURL("http://127.0.0.1:8000/inspector/search/resources/search.js");
        InspectorTest.addResult(resource.url);
        PageAgent.searchInResource(resource.frameId, resource.url, text, step3);
    }

    function step3(error, searchMatches)
    {
        InspectorTest.dumpSearchMatches(searchMatches);
        var resource = WebInspector.resourceForURL("http://127.0.0.1:8000/inspector/search/resources/search.css");
        InspectorTest.addResult(resource.url);
        PageAgent.searchInResource(resource.frameId, resource.url, text, step4);
    }

    function step4(error, searchMatches)
    {
        InspectorTest.dumpSearchMatches(searchMatches);
        var resource = WebInspector.resourceForURL("http://127.0.0.1:8000/inspector/search/resources/search.html");
        InspectorTest.addResult(resource.url);
        PageAgent.searchInResource(resource.frameId, resource.url, text, step5);
    }

    function step5(error, searchMatches)
    {
        InspectorTest.dumpSearchMatches(searchMatches);
        InspectorTest.completeTest();
    }
}
</script>
</head>
<body>
<p>Tests single resource search in inspector page agent.</p>
<a href="https://bugs.webkit.org/show_bug.cgi?id=68998">Bug 68998</a>

<iframe src="resources/search.html" onload="runTest()">
</body>
</html>
